@@ -1,11 +1,12 @@
 {
   "name": "toolz",
-  "version": "3.0.0",
+  "version": "3.1.2",
   "description": "A set of accessible web tools",
   "scripts": {
     "dev": "webpack serve --config config/webpack.dev.js",
     "build": "webpack --config config/webpack.prod.js",
     "build-list": "node src/script/build_list.js",
+    "purge": "node config/purgecss.mjs",
     "format": "prettier --write '**/*.{js,ts}'",
     "lint": "eslint  'src/**/*.{js,ts}' && echo 'Linting complete!'",
     "lint-fix": "eslint  'src/**/*.{js,ts}' --fix"
@@ -34,7 +35,7 @@
     "html-webpack-plugin": "^5.5.3",
     "mini-css-extract-plugin": "^2.7.6",
     "node-sass": "^9.0.0",
-    "postcss": "^8.4.24",
+    "postcss": "^8.4.31",
     "prettier": "^2.8.8",
     "purgecss-webpack-plugin": "^5.0.0",
     "reachable-url": "^1.7.1",
@@ -45,9 +46,9 @@
     "webpack-merge": "^5.9.0"
   },
   "devDependencies": {
-    "ejs": "^3.1.9",
+    "ejs": "^3.1.10",
     "template-ejs-loader": "^0.9.4",
-    "webpack": "^5.88.0",
+    "webpack": "^5.95.0",
     "webpack-cli": "^5.1.4"
   }
 }
