@@ -1,247 +1,247 @@
 export const font_list = [
-    "Abadi MT Condensed Light",
-    "Albertus Extra Bold",
-    "Albertus Medium",
-    "Amazone BT",
-    "AmerType Md BT",
-    "American Typewriter",
-    "Andale Mono",
-    "Antique Olive",
-    "Arial",
-    "Arial Black",
-    "Arial MT",
-    "Arial Narrow",
-    "Arial Rounded MT Bold",
-    "Arial Unicode MS",
-    "Arrus BT",
-    "Aurora Cn BT",
-    "AvantGarde Bk BT",
-    "AvantGarde Md BT",
-    "Avenir",
-    "Avenir Next",
-    "Avenir Next Condensed",
-    "Bahnschrift",
-    "BankGothic Md BT",
-    "Baskerville",
-    "Bazooka",
-    "Benguiat Bk BT",
-    "BernhardFashion BT",
-    "BernhardMod BT",
-    "Big Caslon",
-    "BinnerD",
-    "Bodoni 72",
-    "Bodoni 72 Oldstyle",
-    "Bodoni 72 Smallcaps",
-    "Book Antiqua",
-    "Bookman Old Style",
-    "Boulder",
-    "Bradley Hand",
-    "Bremen Bd BT",
-    "Brush Script MT",
-    "CG Omega",
-    "CG Times",
-    "Calibri",
-    "Calisto MT",
-    "Calligrapher",
-    "Cambria",
-    "Cambria Math",
-    "Candara",
-    "CaslonOpnface BT",
-    "Century Gothic",
-    "Century Schoolbook",
-    "Cezanne",
-    "Chalkboard",
-    "Chalkboard SE",
-    "Chalkduster",
-    "Charlesworth",
-    "Charter",
-    "Charter BT",
-    "Charter Bd BT",
-    "Chaucer",
-    "ChelthmITC Bk BT",
-    "Clarendon Condensed",
-    "CloisterBlack BT",
-    "Cochin",
-    "Comic Sans MS",
-    "Consolas",
-    "Constantia",
-    "CopperplGoth Bd BT",
-    "Copperplate",
-    "Copperplate Gothic Bold",
-    "Copperplate Gothic Light",
-    "Corbel",
-    "Cornerstone",
-    "Coronet",
-    "Courier",
-    "Courier New",
-    "Cuckoo",
-    "DIN Alternate",
-    "DIN Condensed",
-    "Dauphin",
-    "Denmark",
-    "Didot",
-    "Ebrima",
-    "English 111 Vivace BT",
-    "EngraversGothic BT",
-    "Exotc350 Bd BT",
-    "Franklin Gothic Medium",
-    "Fransiscan",
-    "Freefrm721 Blk BT",
-    "FrnkGothITC Bk BT",
-    "Futura",
-    "Futura Bk BT",
-    "Futura Lt BT",
-    "Futura Md BT",
-    "Futura ZBlk BT",
-    "FuturaBlack BT",
-    "Gabriola",
-    "Gadugi",
-    "Galliard BT",
-    "Garamond",
-    "Geneva",
-    "GeoSlab 703 Lt BT",
-    "GeoSlab 703 XBd BT",
-    "Geometr231 BT",
-    "Geometr231 Hv BT",
-    "Geometr231 Lt BT",
-    "Georgia",
-    "Gill Sans",
-    "GoudyHandtooled BT",
-    "GoudyOLSt BT",
-    "Haettenschweiler",
-    "Heather",
-    "Helvetica",
-    "Helvetica Neue",
-    "Herald",
-    "Herculanum",
-    "Hoefler Text",
-    "HoloLens MDL2 Assets",
-    "Humanst 521 Cn BT",
-    "Humanst521 BT",
-    "Humanst521 Lt BT",
-    "Impact",
-    "Incised901 BT",
-    "Incised901 Bd BT",
-    "Incised901 Lt BT",
-    "Informal011 BT",
-    "Ink Free",
-    "Javanese Text",
-    "Jester",
-    "Kabel Bk BT",
-    "Kabel Ult BT",
-    "Kaufmann BT",
-    "Kaufmann Bd BT",
-    "Korinna BT",
-    "Leelawadee UI",
-    "Letter Gothic",
-    "Lithograph",
-    "Lithograph Light",
-    "Long Island",
-    "Lucida Console",
-    "Lucida Grande",
-    "Lucida Handwriting",
-    "Lucida Sans",
-    "Lucida Sans Unicode",
-    "Luminari",
-    "Lydian BT",
-    "MS Gothic",
-    "MS LineDraw",
-    "MV Boli",
-    "Malgun Gothic",
-    "Marigold",
-    "Marker Felt",
-    "Market",
-    "Marlett",
-    "Matisse ITC",
-    "Menlo",
-    "Microsoft Himalaya",
-    "Microsoft JhengHei",
-    "Microsoft New Tai Lue",
-    "Microsoft PhagsPa",
-    "Microsoft Sans Serif",
-    "Microsoft Tai Le",
-    "Microsoft YaHei",
-    "Microsoft Yi Baiti",
-    "MingLiU-ExtB",
-    "Monaco",
-    "Mongolian Baiti",
-    "Monotype Corsiva",
-    "Myanmar Text",
-    "News GothicMT",
-    "NewsGoth BT",
-    "Nirmala UI",
-    "Noteworthy",
-    "OCR A Extended",
-    "Old Century",
-    "Onyx BT",
-    "Optima",
-    "OzHandicraft BT",
-    "PTBarnum BT",
-    "Palatino",
-    "Palatino Linotype",
-    "Papyrus",
-    "Pegasus",
-    "Phosphate",
-    "Pickwick",
-    "Poster",
-    "PosterBodoni BT",
-    "Pythagoras",
-    "Ribbon131 Bd BT",
-    "Rockwell",
-    "Savoye LET",
-    "Sceptre",
-    "Segoe MDL2 Assets",
-    "Segoe Print",
-    "Segoe Script",
-    "Segoe UI",
-    "Segoe UI Emoji",
-    "Segoe UI Historic",
-    "Segoe UI Symbol",
-    "Serifa BT",
-    "Serifa Th BT",
-    "ShelleyVolante BT",
-    "Sherwood",
-    "SignPainter",
-    "Signboard",
-    "SimSun",
-    "Sitka",
-    "Skia",
-    "Snell Roundhand",
-    "Socket",
-    "Souvenir Lt BT",
-    "Staccato222 BT",
-    "Steamer",
-    "Storybook",
-    "Subway",
-    "Swis721 BlkEx BT",
-    "Swiss911 XCm BT",
-    "Sylfaen",
-    "Symbol",
-    "Tahoma",
-    "Technical",
-    "Teletype",
-    "Tempus Sans ITC",
-    "Times",
-    "Times New Roman",
-    "Times New Roman PS",
-    "Trattatello",
-    "Trebuchet MS",
-    "Tristan",
-    "Tubular",
-    "TypoUpright BT",
-    "Unicorn",
-    "Univers",
-    "Univers Condensed",
-    "Vagabond",
-    "Verdana",
-    "Webdings",
-    "Westminster\tAllegro",
-    "Wingdings",
-    "Yu Gothic",
-    "ZapfEllipt BT",
-    "ZapfHumnst BT",
-    "ZapfHumnst Dm BT",
-    "Zapfino",
-    "Zurich BlkEx BT",
-    "Zurich Ex BT"
+	'Abadi MT Condensed Light',
+	'Albertus Extra Bold',
+	'Albertus Medium',
+	'Amazone BT',
+	'AmerType Md BT',
+	'American Typewriter',
+	'Andale Mono',
+	'Antique Olive',
+	'Arial',
+	'Arial Black',
+	'Arial MT',
+	'Arial Narrow',
+	'Arial Rounded MT Bold',
+	'Arial Unicode MS',
+	'Arrus BT',
+	'Aurora Cn BT',
+	'AvantGarde Bk BT',
+	'AvantGarde Md BT',
+	'Avenir',
+	'Avenir Next',
+	'Avenir Next Condensed',
+	'Bahnschrift',
+	'BankGothic Md BT',
+	'Baskerville',
+	'Bazooka',
+	'Benguiat Bk BT',
+	'BernhardFashion BT',
+	'BernhardMod BT',
+	'Big Caslon',
+	'BinnerD',
+	'Bodoni 72',
+	'Bodoni 72 Oldstyle',
+	'Bodoni 72 Smallcaps',
+	'Book Antiqua',
+	'Bookman Old Style',
+	'Boulder',
+	'Bradley Hand',
+	'Bremen Bd BT',
+	'Brush Script MT',
+	'CG Omega',
+	'CG Times',
+	'Calibri',
+	'Calisto MT',
+	'Calligrapher',
+	'Cambria',
+	'Cambria Math',
+	'Candara',
+	'CaslonOpnface BT',
+	'Century Gothic',
+	'Century Schoolbook',
+	'Cezanne',
+	'Chalkboard',
+	'Chalkboard SE',
+	'Chalkduster',
+	'Charlesworth',
+	'Charter',
+	'Charter BT',
+	'Charter Bd BT',
+	'Chaucer',
+	'ChelthmITC Bk BT',
+	'Clarendon Condensed',
+	'CloisterBlack BT',
+	'Cochin',
+	'Comic Sans MS',
+	'Consolas',
+	'Constantia',
+	'CopperplGoth Bd BT',
+	'Copperplate',
+	'Copperplate Gothic Bold',
+	'Copperplate Gothic Light',
+	'Corbel',
+	'Cornerstone',
+	'Coronet',
+	'Courier',
+	'Courier New',
+	'Cuckoo',
+	'DIN Alternate',
+	'DIN Condensed',
+	'Dauphin',
+	'Denmark',
+	'Didot',
+	'Ebrima',
+	'English 111 Vivace BT',
+	'EngraversGothic BT',
+	'Exotc350 Bd BT',
+	'Franklin Gothic Medium',
+	'Fransiscan',
+	'Freefrm721 Blk BT',
+	'FrnkGothITC Bk BT',
+	'Futura',
+	'Futura Bk BT',
+	'Futura Lt BT',
+	'Futura Md BT',
+	'Futura ZBlk BT',
+	'FuturaBlack BT',
+	'Gabriola',
+	'Gadugi',
+	'Galliard BT',
+	'Garamond',
+	'Geneva',
+	'GeoSlab 703 Lt BT',
+	'GeoSlab 703 XBd BT',
+	'Geometr231 BT',
+	'Geometr231 Hv BT',
+	'Geometr231 Lt BT',
+	'Georgia',
+	'Gill Sans',
+	'GoudyHandtooled BT',
+	'GoudyOLSt BT',
+	'Haettenschweiler',
+	'Heather',
+	'Helvetica',
+	'Helvetica Neue',
+	'Herald',
+	'Herculanum',
+	'Hoefler Text',
+	'HoloLens MDL2 Assets',
+	'Humanst 521 Cn BT',
+	'Humanst521 BT',
+	'Humanst521 Lt BT',
+	'Impact',
+	'Incised901 BT',
+	'Incised901 Bd BT',
+	'Incised901 Lt BT',
+	'Informal011 BT',
+	'Ink Free',
+	'Javanese Text',
+	'Jester',
+	'Kabel Bk BT',
+	'Kabel Ult BT',
+	'Kaufmann BT',
+	'Kaufmann Bd BT',
+	'Korinna BT',
+	'Leelawadee UI',
+	'Letter Gothic',
+	'Lithograph',
+	'Lithograph Light',
+	'Long Island',
+	'Lucida Console',
+	'Lucida Grande',
+	'Lucida Handwriting',
+	'Lucida Sans',
+	'Lucida Sans Unicode',
+	'Luminari',
+	'Lydian BT',
+	'MS Gothic',
+	'MS LineDraw',
+	'MV Boli',
+	'Malgun Gothic',
+	'Marigold',
+	'Marker Felt',
+	'Market',
+	'Marlett',
+	'Matisse ITC',
+	'Menlo',
+	'Microsoft Himalaya',
+	'Microsoft JhengHei',
+	'Microsoft New Tai Lue',
+	'Microsoft PhagsPa',
+	'Microsoft Sans Serif',
+	'Microsoft Tai Le',
+	'Microsoft YaHei',
+	'Microsoft Yi Baiti',
+	'MingLiU-ExtB',
+	'Monaco',
+	'Mongolian Baiti',
+	'Monotype Corsiva',
+	'Myanmar Text',
+	'News GothicMT',
+	'NewsGoth BT',
+	'Nirmala UI',
+	'Noteworthy',
+	'OCR A Extended',
+	'Old Century',
+	'Onyx BT',
+	'Optima',
+	'OzHandicraft BT',
+	'PTBarnum BT',
+	'Palatino',
+	'Palatino Linotype',
+	'Papyrus',
+	'Pegasus',
+	'Phosphate',
+	'Pickwick',
+	'Poster',
+	'PosterBodoni BT',
+	'Pythagoras',
+	'Ribbon131 Bd BT',
+	'Rockwell',
+	'Savoye LET',
+	'Sceptre',
+	'Segoe MDL2 Assets',
+	'Segoe Print',
+	'Segoe Script',
+	'Segoe UI',
+	'Segoe UI Emoji',
+	'Segoe UI Historic',
+	'Segoe UI Symbol',
+	'Serifa BT',
+	'Serifa Th BT',
+	'ShelleyVolante BT',
+	'Sherwood',
+	'SignPainter',
+	'Signboard',
+	'SimSun',
+	'Sitka',
+	'Skia',
+	'Snell Roundhand',
+	'Socket',
+	'Souvenir Lt BT',
+	'Staccato222 BT',
+	'Steamer',
+	'Storybook',
+	'Subway',
+	'Swis721 BlkEx BT',
+	'Swiss911 XCm BT',
+	'Sylfaen',
+	'Symbol',
+	'Tahoma',
+	'Technical',
+	'Teletype',
+	'Tempus Sans ITC',
+	'Times',
+	'Times New Roman',
+	'Times New Roman PS',
+	'Trattatello',
+	'Trebuchet MS',
+	'Tristan',
+	'Tubular',
+	'TypoUpright BT',
+	'Unicorn',
+	'Univers',
+	'Univers Condensed',
+	'Vagabond',
+	'Verdana',
+	'Webdings',
+	'Westminster\tAllegro',
+	'Wingdings',
+	'Yu Gothic',
+	'ZapfEllipt BT',
+	'ZapfHumnst BT',
+	'ZapfHumnst Dm BT',
+	'Zapfino',
+	'Zurich BlkEx BT',
+	'Zurich Ex BT'
 ]
