@@ -16,7 +16,7 @@ if (tzversion !== version) {
 	//Show changelog
 	//ch_dialog.show()
 	//Set version
-	TZ.set('version', version)
+	//TZ.set('version', version)
 }
 // Call the function when the DOM is loaded
 document.addEventListener('DOMContentLoaded', () => {
